Upgraded to Production Level project

+++ BackEnd/routes/order_routes.py
@@ -11,10 +11,11 @@ order_bp = Blueprint('order_bp', __name__, url_prefix='/order')
 def add_to_order():
     user_id = int(get_jwt_identity())
     data = request.json
-    product_id = data.get('product_id')
+    product_id = data.get('id')
     quantity = data.get('quantity', 1)
 
     product = Products.query.get(product_id)
+    print(data)
     if not product:
         return jsonify({'status': 'Product not found'}), 404
 
@@ -36,15 +37,19 @@ def add_to_order():
 def fetch_orders():
     user_id = int(get_jwt_identity())
     orders = Order_T.query.filter_by(user_id=user_id).all()
+    
     return jsonify([order.to_dict() for order in orders]), 200
 
 
 @order_bp.route('/delete/<int:item_id>', methods=['DELETE'])
 @jwt_required()
 def delete_order_item(item_id):
+    user_id = int(get_jwt_identity())
     item = Order_T.query.get(item_id)
     if not item:
         return jsonify({'status': 'Item not found'}), 404
+    if item.user_id != user_id:
+        return jsonify({'status': 'Unauthorized'}), 403
 
     db.session.delete(item)
     db.session.commit()
@@ -54,6 +59,7 @@ def delete_order_item(item_id):
 @order_bp.route('/update/<int:item_id>', methods=['PUT'])
 @jwt_required()
 def update_quantity(item_id):
+    user_id = int(get_jwt_identity())
     data = request.json
     new_qty = data.get('quantity')
     if new_qty is None or new_qty < 1:
@@ -62,7 +68,8 @@ def update_quantity(item_id):
     order_item = Order_T.query.get(item_id)
     if not order_item:
         return jsonify({'status': 'Item not found'}), 404
-
+    if order_item.user_id != user_id:
+        return jsonify({'status': 'Unauthorized'}), 403
     order_item.quantity = new_qty
     db.session.commit()
     return jsonify({'status': 'Quantity updated successfully'}), 200
